# Style and performance improvements