Best reviewed commit-wise.

<s>

## Commit 1

The first commit changes the process for changing variable bounds from one that made a few vectors, looped through data <s>twice</s>thrice, and then made two C calls, to a simple loop that sets the bounds element-wise. For small numbers of variables (e.g., changing the bound of a single variable), this PR comes out ahead. For moderate numbers (e.g, 10 variables), it is ~5% slower, and then for large numbers (e.g., > 100) it's back in front. I must have tested on different LQOI commits. Here are the latest benchmarks:

```julia
using Gurobi, BenchmarkTools
const MOI = Gurobi.MOI

function setup(N)
    model = Gurobi.Optimizer()
    x = MOI.add_variables(model, N)
    c = MOI.add_constraints(model, MOI.SingleVariable.(x), Ref(MOI.EqualTo(1.0)))
    return model, c
end

function bench(model, constraints)
    for constraint in constraints
        MOI.set(model, MOI.ConstraintSet(), constraint, MOI.EqualTo(2.0))
    end
end
```

### Before

```julia
julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^0))
2.600 μs (14 allocations: 1.08 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^1))
25.042 μs (140 allocations: 10.78 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^2))
271.362 μs (1400 allocations: 107.81 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^3))
6.075 ms (14978 allocations: 1.07 MiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^4))
427.902 ms (158978 allocations: 10.82 MiB)
```

### After

```julia
julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^0))
  1.437 μs (6 allocations: 336 bytes)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^1))
  13.958 μs (60 allocations: 3.28 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^2))
  234.004 μs (600 allocations: 32.81 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^3))
  5.629 ms (6978 allocations: 343.41 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^4))
  431.477 ms (78978 allocations: 3.49 MiB)
```
</s>

Things slow down for more than 10^4 variables. We really need to be able to set multiple constraints sets simultaneously...

## Commit 2

Instead of querying the full rays, we just try to access the first element. This results in a 2x improvement. I looked at changing from a try-catch to an explicit C call to get the return value, but it was a pretty minor improvement for extra low-level code lying around.

## Commit 3

No real performance difference, but removes 1 allocation each time we query the status.

## Commit 4

Probably used infrequently, but it seemed pointless to perform a copy here rather than switching from 0-indexing to 1-indexing in-place.

## Commit 5

Style updates.