@joaquimg 2018-11-06
I have a few questions about point 1. 
2 to 5 look great!