@mlubin 2018-11-08
```
julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^2))
  234.004 μs (600 allocations: 32.81 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^3))
  5.629 ms (6978 allocations: 343.41 KiB)

julia> @btime bench(model, constraints) setup=((model, constraints) = setup(10^4))
  431.477 ms (78978 allocations: 3.49 MiB)
```

Why doesn't the setup time grow linearly with the size of the problem? What are we doing wrong?