@odow 2018-11-08
This will teach me for benchmarking the wrong function... I should have timed `set_variable_bounds` explicitly. The majority of the time is in LQOI which gives me a different place to go hunting. Re our email discussion earlier, I'm getting confused with all these layers...

```julia
using Gurobi, BenchmarkTools
const MOI = Gurobi.MOI
const LQOI = Gurobi.LQOI

function constr_setup2(N)
    model = Gurobi.Optimizer()
    x = MOI.add_variables(model, N)
    sets = fill(MOI.EqualTo(1.0), N)
    return model, MOI.SingleVariable.(x), sets
end
function constr_run2(model, variables, sets)
    LQOI.set_variable_bounds(model, variables, sets)
end

@btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^1));
@btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^2));
@btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^3));
@btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^4));
```

With commit 1:
```julia
julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^1));
  7.636 μs (8 allocations: 1.09 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^2));
  100.170 μs (8 allocations: 6.00 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^3));
  901.529 μs (8 allocations: 51.64 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^4));
  9.004 ms (13 allocations: 508.59 KiB)
```

After reverting commit 1:
```julia
julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^1));
  2.691 μs (16 allocations: 2.28 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^2));
  8.211 μs (16 allocations: 12.31 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^3));
  94.833 μs (16 allocations: 107.39 KiB)

julia> @btime constr_run2(m, v, s) setup = ((m, v, s) = constr_setup2(10^4));
  1.062 ms (29 allocations: 1.03 MiB)
```