@odow 2018-11-08
The setting of variable bounds could be improved, but this PR is not the place. I've just dropped the commit. Now this PR just contains minor, in-offensive updates.